Fix #230: Changed float to int

+++ friture/generators/sweep.py
@@ -155,9 +155,9 @@ class SettingsWidget(QtWidgets.QWidget):
         settings.setValue("sweep period", self.spinBox_sweep_period.value())
 
     def restoreState(self, settings):
-        sweep_start_frequency = settings.value("sweep start frequency", DEFAULT_SWEEP_STARTFREQUENCY, type=float)
+        sweep_start_frequency = settings.value("sweep start frequency", DEFAULT_SWEEP_STARTFREQUENCY, type=int)
         self.spinBox_sweep_startfrequency.setValue(sweep_start_frequency)
-        sweep_stop_frequency = settings.value("sweep stop frequency", DEFAULT_SWEEP_STOPFREQUENCY, type=float)
+        sweep_stop_frequency = settings.value("sweep stop frequency", DEFAULT_SWEEP_STOPFREQUENCY, type=int)
         self.spinBox_sweep_stopfrequency.setValue(sweep_stop_frequency)
         sweep_period = settings.value("sweep period", DEFAULT_SWEEP_PERIOD_S, type=float)
         self.spinBox_sweep_period.setValue(sweep_period)
